pyproject.toml
[project]
name = "chess-puzzle-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=3.4.3",
    "fire>=0.7.0",
    "jupyterlab>=4.4.3",
    "logfire>=3.21.1",
    "matplotlib>=3.10.3",
    "mlflow>=3.1.0",
    "openrouter>=1.0",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "pydantic-ai>=0.3.1",
    "python-chess>=1.999",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pdbpp>=0.11.6",
    "ruff>=0.12.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
